Split threat models
